ref(replay): Refactor replay for mobile touch start/end parity check
